Some corerction for removng warnings

+++ test/cl/ReduceByKeyTest/ReduceByKeyTest.cpp
@@ -251,9 +251,9 @@ gold_reduce_by_key( InputIterator1 keys_first,
         refInput[i] = i;
     }
 
-	bolt::cl::device_vector<int>  input( refInput.begin(), refInput.end() );
-	bolt::cl::device_vector<int> device_keys(keys.begin(), keys.end());
-	
+    bolt::cl::device_vector<int>  input( refInput.begin(), refInput.end() );
+    bolt::cl::device_vector<int> device_keys(keys.begin(), keys.end());
+    
     // input and output vectors for device and reference
 
     bolt::cl::device_vector<int>  koutput( length );
@@ -305,8 +305,8 @@ gold_reduce_by_key( InputIterator1 keys_first,
     }
 
     bolt::cl::device_vector<int>  input( refInput.begin(), refInput.end() );
-	bolt::cl::device_vector<int> device_keys(keys.begin(), keys.end());
-	
+    bolt::cl::device_vector<int> device_keys(keys.begin(), keys.end());
+    
     // input and output vectors for device and reference
 
     bolt::cl::device_vector<int>  koutput( length );
@@ -363,8 +363,8 @@ gold_reduce_by_key( InputIterator1 keys_first,
     // input and output vectors for device and reference
 
     bolt::cl::device_vector<int>  input( refInput.begin(), refInput.end() );
-	bolt::cl::device_vector<int> device_keys(keys.begin(), keys.end());
-	
+    bolt::cl::device_vector<int> device_keys(keys.begin(), keys.end());
+    
     bolt::cl::device_vector<int>  koutput( length );
     bolt::cl::device_vector<int>  voutput( length );
     std::vector<int>  krefOutput( length );
@@ -402,11 +402,11 @@ for (int i = 0; i < std::ceil(size/3.0); i++)
 {
     vectKeyIn[i] = (int)2;
 }
-for (int i = (std::ceil(size/3.0) + 1); i < std::ceil((2* size)/3.0); i++)
+for (int i = int(std::ceil(size/3.0) + 1); i < std::ceil((2* size)/3.0); i++)
 {
 vectKeyIn[i] = (int)3;
 }
-for (int i =  (std::ceil((2* size)/3.0) + 1); i < size; i++)
+for (int i =  int(std::ceil((2* size)/3.0) + 1); i < size; i++)
 {
 vectKeyIn[i] = (int)5;
 }
@@ -440,11 +440,11 @@ for (int i = 0; i < std::ceil(size/3.0); i++)
 {
     vectKeyIn[i] = (int)2;
 }
-for (int i = (std::ceil(size/3.0) + 1); i < std::ceil((2* size)/3.0); i++)
+for (int i = int(std::ceil(size/3.0) + 1); i < std::ceil((2* size)/3.0); i++)
 {
 vectKeyIn[i] = (int)3;
 }
-for (int i =  (std::ceil((2* size)/3.0) + 1); i < size; i++)
+for (int i =  int(std::ceil((2* size)/3.0) + 1); i < size; i++)
 {
 vectKeyIn[i] = (int)5;
 }
@@ -481,11 +481,11 @@ for (int i = 0; i < std::ceil(size/3.0); i++)
 {
     vectKeyIn[i] = (int)2;
 }
-for (int i = (std::ceil(size/3.0) + 1); i < std::ceil((2* size)/3.0); i++)
+for (int i = int(std::ceil(size/3.0) + 1); i < std::ceil((2* size)/3.0); i++)
 {
 vectKeyIn[i] = (int)3;
 }
-for (int i =  (std::ceil((2* size)/3.0) + 1); i < size; i++)
+for (int i =  int(std::ceil((2* size)/3.0) + 1); i < size; i++)
 {
 vectKeyIn[i] = (int)5;
 }
